refactor: move formatting out of the layers that should not own it
docs/architecture.md の層分離表に照らして2つずれていた。

rules/skill-roll.ts が utils/helper.ts の formatDMPart を import しており、
rules/ から外へ出る唯一の依存になっていた。「2DM≦6」「(2+6)DM≦(5-2)」という
式の見せ方は判定ルールの一部なので、skill-roll.ts に取り込む。utils/helper.ts は
これしか持っておらず名前も汎用すぎたので、テストごと rules/ へ移して削除した。
これで rules/ は相対importとvitestだけになり、完全に自己完結する。

documents/actor.ts の formatSkillName は「＊格闘」「★技能：専門」を組み立てる
表示整形で、表では documents/ は薄いオーケストレーションと決めている。チャット
カードの見せ方なので utils/chat.ts の createRollMessage の隣へ移した。

移動のついでに actor.ts の SkillRollContext のimportが未使用になっていたので
外した。なお biome の noUnusedImports は警告どまりで biome ci が終了コード0を
返すため、これは lint では捕まらない。tsc の noUnusedLocals なら捕まえられる
ことを確認したので、別途そちらで塞ぐ。

判定名の4パターン（通常・特化・エクストラ・基本技能）と共鳴を実機で確認し、
それぞれ〈技巧〉〈専門知識：考古学〉〈★強運〉〈＊手当て〉〈♾️共鳴〉と
これまでどおり出ることを確かめた。エラー・通知ともゼロ。
