+++ module/rules/skill-roll.ts
@@ -1,4 +1,3 @@
-import { formatDMPart } from "../utils/helper";
 import type { ModifierSet, RollSpec } from "./types";
 
 export type SkillRollParams = {
@@ -34,4 +33,16 @@ export function resolveSkillRoll({
     successMod,
     dmFormula: `${formatDMPart(level, bonus)}DM≦${formatDMPart(baseTarget, targetMod)}`,
   };
+}
+
+/**
+ * DM式の一項を組み立てる。修正がなければ基準値だけ、あれば括弧でくくって符号を添える。
+ *
+ * 「2DM≦6」「(2+6)DM≦(5-2)」のように、判定の内訳が読み取れる形にするためのもの。
+ * 式の見せ方は判定ルールの一部なのでこの層に置く。
+ */
+export function formatDMPart(base: number, modifier: number): string {
+  if (!modifier) return `${base}`;
+  const sign = modifier > 0 ? `+${modifier}` : modifier;
+  return `(${base}${sign})`;
 }
