+++ module/utils/chat.ts
@@ -1,3 +1,4 @@
+import type { SkillRollContext } from "../data/character";
 import type { EmokloreRoll } from "../dice/emoklore-roll";
 import type { EmokloreActor } from "../documents/actor";
 
@@ -27,4 +28,22 @@ export async function createRollMessage({
   });
 
   return created as ChatMessage | undefined;
+}
+
+/**
+ * チャットの見出しに出す判定名を組み立てる。「＊格闘」「★技能：専門」など。
+ *
+ * 基本技能は「＊」、エクストラ技能は「★」を頭に付けるのがシート表記の慣習。
+ * 表示の話なのでルール層ではなく、チャットカードを作るこの層に置く。
+ */
+export function formatSkillName(
+  { label, isExtra, specialization }: SkillRollContext,
+  { base }: { base: boolean },
+): string {
+  const prefix = base ? "＊" : isExtra ? "★" : "";
+  const suffix = specialization
+    ? `${game.i18n.localize("EMOKLORE.Common.colon")}${specialization}`
+    : "";
+
+  return `${prefix}${label}${suffix}`;
 }
