--- module/utils/helper.ts
@@ -1,5 +1,0 @@
-export function formatDMPart(base: number, modifier: number): string {
-  if (!modifier) return `${base}`;
-  const sign = modifier > 0 ? `+${modifier}` : modifier;
-  return `(${base}${sign})`;
-}
